requirements.txt
Flask
Flask-SQLAlchemy
Flask-RESTful
openai
python-dotenv